+++ web/documentserver-example/python/src/utils/users.py
@@ -19,7 +19,8 @@
 from typing import Optional
 
 class User:
-    def __init__(self, id, name, email, group, reviewGroups, commentGroups, userInfoGroups, favorite, deniedPermissions, descriptions, templates):
+    def __init__(self, id, name, email, group, reviewGroups, commentGroups, userInfoGroups, favorite,
+                 deniedPermissions, descriptions, templates, avatar):
         self.id = id
         self.name = name
         self.email = email
@@ -31,6 +32,7 @@ class User:
         self.descriptions = descriptions
         self.templates = templates
         self.userInfoGroups = userInfoGroups
+        self.avatar = avatar
 
 descr_user_1 = [
     "File author by default",
@@ -39,7 +41,8 @@ descr_user_1 = [
     "Can perform all actions with comments",
     "The file favorite state is undefined",
     "Can create files from templates using data from the editor",
-    "Can see the information about all users"
+    "Can see the information about all users",
+    "Has an avatar"
 ]
 
 descr_user_2 = [
@@ -48,7 +51,8 @@ descr_user_2 = [
     "Can view comments, edit his own comments and comments left by users with no group. Can remove his own comments only",
     "This file is marked as favorite",
     "Can create new files from the editor",
-    "Can see the information about users from Group2 and users who don’t belong to any group"
+    "Can see the information about users from Group2 and users who don’t belong to any group",
+    "Has an avatar"
 ]
 
 descr_user_3 = [
@@ -81,7 +85,7 @@ descr_user_0 = [
 USERS = [
     User('uid-1', 'John Smith', 'smith@example.com',
         '', None, {}, None,
-        None, [], descr_user_1, True),
+        None, [], descr_user_1, True, True),
     User('uid-2', 'Mark Pottato', 'pottato@example.com',
         'group-2', ['group-2', ''], {
             'view': "",
@@ -89,17 +93,17 @@ USERS = [
             'remove': ["group-2"]
         },
          ['group-2', ''],
-        True, [], descr_user_2, False),
+        True, [], descr_user_2, False, True),
     User('uid-3', 'Hamish Mitchell', 'mitchell@example.com',
         'group-3', ['group-2'], {
             'view': ["group-3", "group-2"],
             'edit': ["group-2"],
             'remove': []
         }, ['group-2'],
-        False, ["copy", "download", "print"], descr_user_3, False),
+        False, ["copy", "download", "print"], descr_user_3, False, False),
     User('uid-0', None, None,
         '', None, {}, [],
-        None, ["protect"], descr_user_0, False)
+        None, ["protect"], descr_user_0, False, False)
 ]
 
 DEFAULT_USER = USERS[0]