+++ web/documentserver-example/python/src/views/actions.py
@@ -211,6 +211,16 @@ def edit(request):
         }
     ]
 
+    usersInfo = []
+    if user.id != 'uid-0':
+        for userInfo in users.getAllUsers():
+            u = userInfo
+            u.image = docManager.getServerUrl(True, request) + f'/static/images/{u.id}.jpg' if user.avatar else None
+            usersInfo.append({"id": u.id, "name": u.name, "email": u.email, "image": u.image, "group": u.group,
+                              "reviewGroups":u.reviewGroups, "commentGroups":u.commentGroups, "favorite": u.favorite,
+                              "deniedPermissions": u.deniedPermissions, "descriptions": u.descriptions,
+                              "templates": u.templates, "userInfoGroups": u.userInfoGroups, "avatar": u.avatar})
+
     if (meta):  # if the document meta data exists,
         infObj = {  # write author and creation time parameters to the information object
             'owner': meta['uname'],
@@ -269,7 +279,8 @@ def edit(request):
             'user': {  # the user currently viewing or editing the document
                 'id': user.id if user.id !='uid-0' else None,
                 'name': user.name,
-                'group': user.group
+                'group': user.group,
+                'image': docManager.getServerUrl(True, request) + f'/static/images/{user.id}.jpg' if user.avatar else None
             },
             'embedded': {  # the parameters for the embedded document type
                 'saveUrl': directUrl,  # the absolute URL that will allow the document to be saved onto the user personal computer
@@ -340,7 +351,8 @@ def edit(request):
         'dataInsertImage': json.dumps(dataInsertImage)[1 : len(json.dumps(dataInsertImage)) - 1],  # the image which will be inserted into the document
         'dataDocument': dataDocument,  # document which will be compared with the current document
         'dataSpreadsheet': json.dumps(dataSpreadsheet),  # recipient data for mail merging
-        'usersForMentions': json.dumps(usersForMentions) if user.id !='uid-0' else None
+        'usersForMentions': json.dumps(usersForMentions) if user.id !='uid-0' else None,
+        'usersInfo': json.dumps(usersInfo)
     }
     return render(request, 'editor.html', context)  # execute the "editor.html" template with context data